Some people might be deterred from adopting a parrot because they assume they're all large birds that need a vast amount of space. However, Quakers are just one of the many types of medium-sized birds that prove that notion wrong. Quakers are around 11 to 12 inches long and weigh just 3 to 5 ounces.
While it's true that all birds need as spacious of an enclosure as possible, as well as space outside of the enclosure to play, Quakers and similarly sized species can do well with less space than a large bird, such as a macaw. Make sure your quaker’s housing is sturdy and avian-safe, as these birds are known to chew and to learn how to open cage doors. They also tend to have a strong nest-building instinct. So in addition to toys, offer them bird-safe materials with which they can make a nest if they wish.