Like many parrot species, Quaker parrots can enjoy a particularly long lifespan in captivity. Their average lifespan is between 20 and 30 years. But with quality care, some Quaker parrots can live for more than 30 years.
If you're considering a Quaker, ensure that you have a plan to care for the bird for its entire life. Quakers are a very social species, bonding closely with their owners. So it is difficult for them to handle being bounced from home to home when people have to give them up. In fact, one of the few times Quakers might become aggressive or resort to unhealthy behaviors, such as feather plucking, is when they feel neglected or stressed, which is often the case when they lose their home.